rename: 프로필 관련 패널들을 profile 폴더로 이동
